Rename shuttle-sync-impl to shuttle-std and add thread/future
Merge the sync, thread, and future modules into a single shuttle-std
crate that mirrors std concurrency primitives. The shuttle crate
remains a thin facade that re-exports from shuttle-std.
