+++ Cargo.toml
@@ -2,7 +2,7 @@
 members = [
     "shuttle-core",
     "shuttle-schedulers",
-    "shuttle-sync-impl",
+    "shuttle-std",
     "shuttle",
     "wrappers/shuttle_rand_0.8",
     "wrappers/shuttle_sync",
