+++ shuttle-std/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "shuttle-sync-impl"
+name = "shuttle-std"
 version = "0.1.0"
 edition = "2021"
 license = "Apache-2.0"
-description = "Shuttle implementations of std::sync primitives"
+description = "Shuttle implementations of std synchronization, threading, and async primitives"
 repository = "https://github.com/awslabs/shuttle"
 
 [dependencies]
