+++ shuttle-std/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod future;
+pub mod sync;
+pub mod thread;
