+++ shuttle-std/src/sync/atomic/bool.rs
